+++ examples/connection_templates/update_source.tf
@@ -8,6 +8,7 @@ provider "oneview" {
 }
 
 resource "oneview_connection_templates" "connectionTemplates" {
+  name = "renamed-connection-template"
   bandwidth {
       maximum_bandwidth = 8000
       typical_bandwidth = 2500